Reset project version number to 0.1.0.
This project isn't really semantically versioned per se. It is updated
for new openTCS baseline releases to provide a template that can be used
for new integration projects based on the latest baseline version. Such
new projects should not start with the current baseline version, but
with an initial version number less than 1.0.0.